requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
google-auth==2.39.0
google-genai==1.11.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
PyMuPDF==1.26.0
python-dotenv==1.1.0
python-docx==1.1.2
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1